refactor(pytest-bdd): merge types module into storage
